Cart functions
Fix oversight where including a null for a SelectColumn would try to actually include it in the definition instead of skipping it.
Allow AutoIncrement columns to be optional in TableValidator.
Naive start to PaymentModuleFactory
Augment request attributes for ContactId andEventId into container object for easier usage on objects that don't have access to the Request object.
Allow badges to not have a transaction ID
Add mail template map to specify which mail templates go with which badges