Adjust start/end date handling
Actually use null values instead of not saving changes to the item.
Display date ranges specific badges are available
Allow badges that are not available to be selectable, but not able to be added to the cart (front-end). TODO: Confirm availability when refreshing the cart meta
Group Application addons are a thing we should be supporting...
If a badge type ID is invalid, make that an error in the cart.
Fix some legacy-er carts that didn't have the payment_badge_price in the cart properties from splatting warnings and killing the output.
Working on fixing the badge type editor spewing errors because null is not a valid date
Use the error counts for each cart instead of relying on selected cart state
Request contexts information when trying to view that data in the cart or view badge screens
An attempt to check out cart by UUID if not logged in to the right account (unverified)