Fix deprecation warning of passing NULL
`trim()` expects a string but may potentially be NULL. This caused a
warning and is resolved by using a NULL-COALESCENCE operator.
The same thing goes for `htmlspecialchars()` for example. There are
probably many more of these hiding in the code that should be fixed.
These will be taken care of when they show up.