Fix addon additions breaking the order process
Previously, when the addon purchase was updated in the database but the
`payment_status` was set to `Incomplete` (which is naturally the case)
the `update_addon_purchase_payment_status()` was attempting to set a
`DATETIME` field to the string of the field `payment_status`.

Why. Just why. Somebody had to have tested and found an issue this
elementary, right?! It isn't even hard to run into!