Fix Blacklists not loading
This was caused by turning on warnings on PHP8 which prints deprecation
warnings when passing a `null` value to a function expecting a string
parameter. The blacklists make a call to to blacklist.php expecting a
JSON response in return. The deprecation warnings caused the response to
be invalid when being parsed and prevented the list from loading.
Adding more null coalescence operators resolves this issue by removing
the warnings.