More fixes
Fix missing order clause causing crashes when searching
Fix login confirmation links not having returnTo parameter passed into them
Cause an missing price error to become a more obvious price error to draw attention to it (hopefully).
Allow anons to complete check out of a cart if it's actually ready to be confirmed if they have their UUID (in case they're, for example, returning from a payment on a different device than where the payment was initiated)
Initial work on addons in the admin ui
Work on the add-edit-badge page to hopefully enable/disable the wizard continuation buttons more consistently.