Fixes for admin users
Admin user list now provides more details, if a user has permissions on this event or is a global admin.
When updating, start with the current permissions set, instead of from scratch.
EventPermissions function to check if there are any permissions set at all
Setting an event's permissions with none set will remove it from the permissions blob.
Separate (slightly) the distinction of GlobalAdmin being set at the actual global level or via an event's permission.
Fix some of the lists using a slightly-ugly hack to un-react-ify the apiAddParams object, which was being mutated in the doSearch function