Hotfixes
Checkin screen should pay attention to the payment status as indicated by the intent of the code.
Fix addeditbadge screen to use the correct search method when confirming available add-ons
TODO: Check that already-purchased addons are correctly preserved when changing badge types that do not allow a given addon in the new type.